+++ src/uni_modules/wot-design-uni/components/wd-config-provider/types.ts
@@ -629,16 +629,19 @@ export type tagThemeVars = {
 }
 
 export type toastThemeVars = {
+  toastColor?: string
   toastPadding?: string
   toastMaxWidth?: string
   toastRadius?: string
   toastBg?: string
   toastFs?: string
+  toastLineHeight?: string
   toastWithIconMinWidth?: string
   toastIconSize?: string
   toastIconMarginRight?: string
   toastIconMarginBottom?: string
   toastLoadingPadding?: string
+  toastLoadingMarginBottom?: string
   toastBoxShadow?: string
 }
 
@@ -678,6 +681,8 @@ export type gridItemThemeVars = {
   gridItemBg?: string
   gridItemPadding?: string
   gridItemBorderColor?: string
+  gridItemHoverBg?: string
+  gridItemHoverBgDark?: string
 }
 
 export type statustipThemeVars = {