+++ src/uni_modules/wot-design-uni/components/wd-curtain/wd-curtain.vue
@@ -7,6 +7,7 @@
       :close-on-click-modal="closeOnClickModal"
       :hide-when-close="hideWhenClose"
       :z-index="zIndex"
+      :root-portal="rootPortal"
       @before-enter="beforeenter"
       @enter="enter"
       @after-enter="afterenter"
@@ -85,6 +86,9 @@ watch(
 
 watch(modelValue, (newVal) => {
   emit('update:modelValue', newVal)
+  if (!newVal) {
+    emit('close')
+  }
 })
 
 const imgSucc = ref<boolean>(true)
@@ -125,7 +129,6 @@ function afterleave() {
 
 function close() {
   modelValue.value = false
-  emit('close')
 }
 
 function clickModal() {