+++ src/uni_modules/wot-design-uni/locale/lang/fr-FR.ts
@@ -75,8 +75,8 @@ export default {
   pagination: {
     prev: 'Précédent',
     next: 'Suivant',
-    page: (value: number) => `Page : ${value}`,
-    total: (total: number) => `Total : ${total}`,
+    page: (value: number) => `Page: ${value}`,
+    total: (total: number) => `Total: ${total}`,
     size: (size: number) => `${size}/Page`
   },
   picker: {
@@ -126,6 +126,8 @@ export default {
   },
   signature: {
     confirmText: 'Signer',
-    clearText: 'Effacer'
+    clearText: 'Effacer',
+    revokeText: 'Annuler',
+    restoreText: 'Restaurer'
   }
 }